set up a controller for devise
