+++ config/routes.rb
@@ -2,7 +2,7 @@ Rails.application.routes.draw do
 
   root "challenges#index"
 
-  devise_for :users, :controllers => { registrations: 'registrations' }
+  devise_for :users, :controllers => { registrations: 'users/registrations' }
 
   resources :challenges
